initial overhaul to use Free Music Archive
