+++ .gitignore
@@ -4,3 +4,5 @@
 /.idea/libraries
 .DS_Store
 /build
+/app/manifest-merger-release-report.txt
+/app/app-release.apk
