+++ app/build.gradle
@@ -8,10 +8,12 @@ android {
         applicationId "co.hollowlog.localrhythm"
         minSdkVersion 16
         targetSdkVersion 21
-        versionCode 2
-        versionName "1.1"
+        versionCode 3
+        versionName "2.0"
     }
 }
+
+apply plugin: 'android-unit-test'
 
 repositories {
     mavenCentral()
@@ -24,4 +26,6 @@ dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
     compile 'com.android.support:appcompat-v7:21.0.3'
     compile 'com.spotify.sdk:spotifysdk:1.0.0-beta6@aar'
+    testCompile 'junit:junit:4.10'
+    testCompile 'org.robolectric:robolectric:2.4'
 }
